requirements.txt
PyQt6
requests
Pillow
openai
wordninja
pycryptodome
rubymarshal
pyspellchecker
